refactor: IFS-2400 move oauth client related classes to the oauth.client package
