+++ isy-security/src/main/java/de/bund/bva/isyfact/security/oauth2/client/authentication/IsyOAuth2AuthenticationProvider.java
@@ -1,0 +1,37 @@
+package de.bund.bva.isyfact.security.oauth2.client.authentication;
+
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
+import org.springframework.security.oauth2.client.oidc.authentication.OidcIdTokenDecoderFactory;
+import org.springframework.security.oauth2.client.registration.ClientRegistration;
+import org.springframework.security.oauth2.core.OAuth2AccessToken;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.jwt.JwtDecoderFactory;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
+
+/**
+ * Base class for an AuthenticationProvider that creates a {@link org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken}
+ * from an {@link OAuth2AuthorizedClient}.
+ */
+public abstract class IsyOAuth2AuthenticationProvider implements AuthenticationProvider {
+
+    /** Converter to create a JwtAuthenticationToken from a JWT. */
+    protected final JwtAuthenticationConverter jwtAuthenticationConverter;
+
+    /** Factory for decoding and validating the returned JWT. */
+    private final JwtDecoderFactory<ClientRegistration> jwtDecoderFactory = new OidcIdTokenDecoderFactory();
+
+    public IsyOAuth2AuthenticationProvider(JwtAuthenticationConverter jwtAuthenticationConverter) {
+        this.jwtAuthenticationConverter = jwtAuthenticationConverter;
+    }
+
+    protected Authentication createJwtAuthentication(OAuth2AuthorizedClient authorizedClient) {
+        OAuth2AccessToken accessToken = authorizedClient.getAccessToken();
+        ClientRegistration clientRegistration = authorizedClient.getClientRegistration();
+        Jwt jwt = jwtDecoderFactory.createDecoder(clientRegistration).decode(accessToken.getTokenValue());
+
+        return jwtAuthenticationConverter.convert(jwt);
+    }
+
+}
