+++ isy-security/src/main/java/de/bund/bva/isyfact/security/oauth2/client/authentication/IsyOAuth2ManualClientCredentialsAuthenticationToken.java
@@ -1,4 +1,4 @@
-package de.bund.bva.isyfact.security.authentication;
+package de.bund.bva.isyfact.security.oauth2.client.authentication;
 
 import java.util.Collections;
